chromium/third_party/blink/web_tests/external/wpt/scroll-animations/css/at-scroll-timeline-sampling-expected.txt

This is a testharness.js-based test.
FAIL Scroll position is sampled once per frame assert_equals: expected "0px" but got "100px"
Harness: the test ran to completion.
